+++ example/ansi_escape_codes_example.dart
@@ -4,8 +4,6 @@ import 'dart:io';
 
 import 'package:ansi_escape_codes/ansi.dart';
 import 'package:ansi_escape_codes/ansi_escape_codes.dart';
-import 'package:ansi_escape_codes/extensions.dart';
-import 'package:ansi_escape_codes/utils.dart';
 
 import 'utils.dart';