+++ test/stack_state_test.dart
@@ -1,5 +1,4 @@
 import 'package:ansi_escape_codes/ansi_escape_codes.dart';
-import 'package:ansi_escape_codes/extensions.dart';
 import 'package:ansi_escape_codes/style.dart' as styles;
 import 'package:test/test.dart';